+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2021-11-16
+### Fixed
+- Remove preset operator was not working
+- Adding new preset did not make it active in the preset menu
+
 ## [0.0.2] - 2021-10-14
 ### Changed
 - Code refactoring. Using Daniel Elther method of retrieving keys is more Blender version efficient. Does need manual attributes to be added, works on different version without effort, and thus more efficient
@@ -16,5 +21,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.3]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.3
 [0.0.2]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.2
 [0.0.1]:https://github.com/schroef/Save-Render-Presets/releases/tag/v0.0.1